cpp/bronze/24263번 풀이

+++ cpp/bronze/24263.cpp
@@ -1,7 +1,7 @@
 #include <iostream>
 using namespace std;
 int main() {
-    long long n; cin >> n;
+    int n; cin >> n;
     cout << n << '\n' << 1;
     return 0;
 }
